motors stopped moving after a while.
minor change to arduino robot2.ino, moving marquee.display('o')
    until after the robot synchs.
minor change to udoo/Configurator.h, adding (short) casts to
    the round and part so that it's visible to the screen.

+++ src/Configurator.h
@@ -55,8 +55,8 @@ class Configurator {
 				default:
 					std::cerr << "CONFIG :: acquire --> failure. c_round(" << c_round << ")" << std::endl;
 			}
-			std::cerr << "CONFIG :: acquire --> c_round (" << c_round << ")" << std::endl;
-			std::cerr << "CONFIG :: acquire --> c_part (" << c_part << ")" << std::endl;
+			std::cerr << "CONFIG :: acquire --> c_round (" << (short)c_round << ")" << std::endl;
+			std::cerr << "CONFIG :: acquire --> c_part (" << (short)c_part << ")" << std::endl;
 			std::cerr << "CONFIG :: acquire --> c_start (" << c_start << ")" << std::endl;
 			std::cerr << "CONFIG :: acquire --> c_end (" << c_end << ")" << std::endl;
 		}
